requirements.txt
transformers 
torch 
librosa
SpeechRecognition
soundfile
streamlit
openai 
python-dotenv